methods were added to Parent Page, switchToNewTab was fixed
