+++ DmytryshynProj/src/main/java/org/pages/ParentPage.java
@@ -1,7 +1,10 @@
 package org.pages;
 
+import org.apache.log4j.Logger;
 import org.junit.Assert;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.utils.ConfigProvider;
 
 abstract public class ParentPage extends CommonActionsWithElements{ //create constructor
@@ -10,6 +13,7 @@ abstract public class ParentPage extends CommonActionsWithElements{ //create con
     protected String baseUrl = ConfigProvider.configProperties.base_url().replace("[env]", environment);
 
     public ParentPage(WebDriver webDriver) {super(webDriver);}
+    private Logger logger = Logger.getLogger(getClass());
 
     abstract protected String getRelativeUrl();
 
@@ -28,5 +32,55 @@ abstract public class ParentPage extends CommonActionsWithElements{ //create con
                         "Expected url: " + baseUrl + getRelativeUrl() +
                         "\n Actual url: " + webDriver.getCurrentUrl(),
                 webDriver.getCurrentUrl().matches(baseUrl + getRelativeUrl()));
+    }
+    // open new tab using JS
+    public void openNewTab() {
+        try {
+            ((JavascriptExecutor) webDriver).executeScript("window.open()");
+            logger.info("New tab was opened");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    // switch to new tab
+    public void switchToNewTab(int tabIndex) {
+        try {
+            webDriverWait10.until(ExpectedConditions.numberOfWindowsToBe(tabIndex + 1));
+            webDriver.switchTo().window((String) webDriver.getWindowHandles().toArray()[tabIndex]);
+            logger.info("Switched to tab with index: " + tabIndex);
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    // switch to main tab
+    public void switchToMainTab() {
+        try {
+            webDriver.switchTo().window((String) webDriver.getWindowHandles().toArray()[0]);
+            logger.info("Switched to main tab");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    // close current tab
+    public void closeCurrentTab() {
+        try {
+            webDriver.close();
+            logger.info("Current tab was closed");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    // refresh page
+    public void refreshPage() {
+        try {
+            webDriver.navigate().refresh();
+            logger.info("Page was refreshed");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
     }
 }
